+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 - Add `daps.whitelisted.url` to `application.properties.`. Allows configuring whitelisted DAPS URLs already at configuration time.
 - Introduction of the setting `referred.check=true/false` in the application.properties. Enables or disables DAT claim referringConnector vs IDS message issuerConnector validation. Disabled by default.
 
+### Fixed
+- Fix bootstrapping with broker registration. The DSC now searches for the correct resource id when bootstrapping.
+
 ### Changed
 - Code base refactorings and removing of unused code.